+++ jdaviz/core/events.py
@@ -190,6 +190,30 @@ class AddLineListMessage(Message):
         return self._table
 
 
+class LineIdentifyMessage(Message):
+    def __init__(self, name_rest, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._name_rest = name_rest
+
+    @property
+    def name_rest(self):
+        return self._name_rest
+
+
+class SpectralMarksChangedMessage(Message):
+    def __init__(self, marks, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self._marks = marks
+
+    @property
+    def names_rest(self):
+        return [m.name_rest for m in self.marks]
+
+    @property
+    def marks(self):
+        return self._marks
+
+
 class RedshiftMessage(Message):
     '''Messages related to Specviz redshift slider'''
     def __init__(self, param, value, *args, **kwargs):