+++ jdaviz/core/marks.py
@@ -3,7 +3,7 @@ from bqplot.marks import Lines
 from glue.core import HubListener
 from specutils import Spectrum1D
 
-from jdaviz.core.events import SliceToolStateMessage
+from jdaviz.core.events import SliceToolStateMessage, LineIdentifyMessage
 
 
 class BaseSpectrumVerticalLine(Lines, HubListener):
@@ -57,9 +57,10 @@ class SpectralLine(BaseSpectrumVerticalLine):
     """
     def __init__(self, viewer, rest_value, redshift=0, name=None, **kwargs):
         self._rest_value = rest_value
+        self._identify = False
         self.name = name
 
-        # TODO: do we need table_index anymore?
+        # table_index is same as name_rest elsewhere
         self.table_index = kwargs.pop("table_index", None)
 
         # setting redshift will set self.x and enable the obs_value property,
@@ -68,8 +69,15 @@ class SpectralLine(BaseSpectrumVerticalLine):
         self._x_unit = viewer.state.reference_data.get_object(cls=Spectrum1D).spectral_axis.unit
         self.redshift = redshift
 
+        viewer.session.hub.subscribe(self, LineIdentifyMessage,
+                                     handler=self._process_identify_change)
+
         super().__init__(viewer=viewer, x=self.obs_value, stroke_width=1,
                          fill='none', close_path=False, **kwargs)
+
+    @property
+    def name_rest(self):
+        return self.table_index
 
     @property
     def rest_value(self):
@@ -98,6 +106,21 @@ class SpectralLine(BaseSpectrumVerticalLine):
             obs_value = (obs_angstrom*u.Angstrom).to_value(self._x_unit,
                                                            equivalencies=u.spectral())
         self.x = [obs_value, obs_value]
+
+    @property
+    def identify(self):
+        return self._identify
+
+    @identify.setter
+    def identify(self, identify):
+        if not isinstance(identify, bool):  # pragma: no cover
+            raise TypeError("identify must be of type bool")
+
+        self._identify = identify
+        self.stroke_width = 3 if identify else 1
+
+    def _process_identify_change(self, msg):
+        self.identify = msg.name_rest == self.table_index
 
     def _update_data(self, x_all):
         new_unit = x_all.unit