+++ CHANGES.rst
@@ -4,6 +4,9 @@
 New Features
 ------------
 
+- Lines from the line list plugin can be selected to help identify as well
+  as to assign redshifts from the line analysis plugin. [#1115]
+
 Cubeviz
 ^^^^^^^